+++ flows/actions/base.go
@@ -170,6 +170,14 @@ type universalAction struct{}
 
 // AllowedFlowTypes returns the flow types which this action is allowed to occur in
 func (a *universalAction) AllowedFlowTypes() []flows.FlowType {
+	return []flows.FlowType{flows.FlowTypeMessaging, flows.FlowTypeMessagingBackground, flows.FlowTypeMessagingOffline, flows.FlowTypeVoice}
+}
+
+// utility struct which sets the allowed flow types to non-background
+type interactiveAction struct{}
+
+// AllowedFlowTypes returns the flow types which this action is allowed to occur in
+func (a *interactiveAction) AllowedFlowTypes() []flows.FlowType {
 	return []flows.FlowType{flows.FlowTypeMessaging, flows.FlowTypeMessagingOffline, flows.FlowTypeVoice}
 }
 
@@ -178,7 +186,7 @@ type onlineAction struct{}
 
 // AllowedFlowTypes returns the flow types which this action is allowed to occur in
 func (a *onlineAction) AllowedFlowTypes() []flows.FlowType {
-	return []flows.FlowType{flows.FlowTypeMessaging, flows.FlowTypeVoice}
+	return []flows.FlowType{flows.FlowTypeMessaging, flows.FlowTypeMessagingBackground, flows.FlowTypeVoice}
 }
 
 // utility struct which sets the allowed flow types to just voice