+++ flows/msg.go
@@ -151,6 +151,7 @@ func (m *MsgOut) Topic() MsgTopic { return m.Topic_ }
 type MsgTemplating struct {
 	Template_  *assets.TemplateReference `json:"template"`
 	Language_  envs.Language             `json:"language"`
+	Country_   envs.Country              `json:"country"`
 	Variables_ []string                  `json:"variables,omitempty"`
 }
 
@@ -160,14 +161,18 @@ func (t MsgTemplating) Template() *assets.TemplateReference { return t.Template_
 // Language returns the language that should be used for the template
 func (t MsgTemplating) Language() envs.Language { return t.Language_ }
 
+// Country returns the country that should be used for the template
+func (t MsgTemplating) Country() envs.Country { return t.Country_ }
+
 // Variables returns the variables that should be substituted in the template
 func (t MsgTemplating) Variables() []string { return t.Variables_ }
 
 // NewMsgTemplating creates and returns a new msg template
-func NewMsgTemplating(template *assets.TemplateReference, language envs.Language, variables []string) *MsgTemplating {
+func NewMsgTemplating(template *assets.TemplateReference, language envs.Language, country envs.Country, variables []string) *MsgTemplating {
 	return &MsgTemplating{
 		Template_:  template,
 		Language_:  language,
+		Country_:   country,
 		Variables_: variables,
 	}
 }